Add broken links CI check and fix broken links
This commit adds a .travis.yml file that enables checking the integrity
of all links in README.md in Travis CI at each pull request and
also periodically if we want to. It also adds a Dockerfile and
a Makefile that makes it possible to run the same check locally
in a Docker container.

This commit also fixes a broken link in README.md.